+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Support multiple files of applet in Webpack Plugin
+
+### Fixed
+- Allow CORS in webpack plugin 8090 emulator proxy port for develop applet externally
+
 ## [0.4.4] - 2019-09-24
 ### Fixed
 - Do transpile applet code always with babel-loader to allow run it on any old device out of box